refactor dubbo-samples-stub to springboot style
