+++ 2-advanced/dubbo-samples-stub/dubbo-samples-stub-consumer/src/main/resources/log4j.properties
