+++ 2-advanced/dubbo-samples-stub/dubbo-samples-stub-consumer/src/test/java/org/apache/dubbo/samples/stub/test/ConsumerApplication.java
@@ -1,0 +1,15 @@
+package org.apache.dubbo.samples.stub.test;
+
+import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@EnableDubbo
+public class ConsumerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ConsumerApplication.class, args);
+    }
+}
