+++ 2-advanced/dubbo-samples-stub/dubbo-samples-stub-provider/src/main/resources/log4j.properties
@@ -15,11 +15,12 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 #
+#
 
-
-# Supported component versions of the test case
-
-# Spring app
-dubbo.version=3.*
-spring.version=5.*
-java.version=17
+###set log levels###
+log4j.rootLogger=info, stdout
+###output to the console###
+log4j.appender.stdout=org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.Target=System.out
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern=[%d{dd/MM/yy hh:mm:ss:sss z}] %t %5p %c{2}: %m%n
